+++ src/ReplicatedStorage/OS/HumanoidController/Humanoid.lua
@@ -1,9 +1,9 @@
 local ReplicatedStorage = game:GetService("ReplicatedStorage")
 local ContextActionService = game:GetService("ContextActionService")
-local Rx = require(ReplicatedStorage.OS.Rx)
-local Rxi = require(ReplicatedStorage.OS.Rxi)
+local Rx = require(ReplicatedStorage.Util.Rx)
+local Rxi = require(ReplicatedStorage.Util.Rxi)
 local BaseObject = require(ReplicatedStorage.OS.BaseObject)
-local Maid = require(ReplicatedStorage.OS.Maid)
+local Maid = require(ReplicatedStorage.Util.Maid)
 
 local Humanoid = setmetatable({}, BaseObject)
 Humanoid.__index = Humanoid